Migrate to System.CommandLine 2.0.0-beta7 (#48)
* Add migration plan for System.CommandLine 2.0.0-beta5

* First pass for SCL beta5 migration

* Add command-level validation for subscription option in SubscriptionCommand

* Refactor option definitions to improve readability by moving descriptions to the appropriate property for various Azure tools, including FunctionApp, KeyVault, Kusto, Monitor, MySql, Postgres, Redis, and others.

* Enhance command option handling by normalizing names and improving telemetry assertions in tests

* Refactor command execution validation

- Moved the validation check for command input to the beginning of the ExecuteAsync method in multiple command classes to ensure early return on invalid input.
- Updated the Validate method to handle missing required options more gracefully, providing standardized error messages.
- Introduced helper methods for option normalization and checking presence of option results to streamline validation logic.
- Ensured consistency across various command implementations in the Azure.Mcp.Tools and Azure.Mcp.Core projects.

* Refactor validation checks in command execution methods

- Moved validation checks for command results to the beginning of the ExecuteAsync method in various command classes to streamline error handling and improve code readability.
- Removed redundant validation checks that were previously located deeper within the try block.
- Ensured consistent validation logic across multiple command implementations in the Azure.Mcp.Tools project.

* Refactor command descriptions for consistency and readability across various commands

* Refactor validation logic in ExecuteAsync and Validate methods for improved clarity and consistency

* Refactor command option registration and validation across storage commands

- Updated command option registration to use `command.Options.Add()` instead of `command.AddOption()`.
- Consolidated input validation logic in multiple command classes to ensure consistent validation before executing commands.
- Introduced utility methods for parsing command line arguments and safely retrieving option values from command results.
- Added new test utilities for argument building and splitting to facilitate unit testing.
- Enhanced unit tests to utilize new argument parsing utilities for improved readability and maintainability.

* Refactor test files to improve code consistency and readability

- Updated various test files to ensure consistent indentation and formatting.
- Removed unnecessary using directives and organized imports for clarity.
- Adjusted variable declarations for better readability.
- Ensured that error messages in validation checks are formatted consistently.

* Refactor command option validation and parsing for improved consistency and clarity

* Refactor option registration in FileSystemListCommand and update test cases for improved clarity and consistency

* Skip failing tests in FileSystemListCommandTests and SampleCommandTests for investigation

* Update package version for System.CommandLine and refactor option parsing for consistency

* Remove unused SystemCommandLineExtensions class to improve codebase cleanliness

* Refactor command files to remove unused System.CommandLine references and update parsing methods for consistency

* Clarify comment in Validate method to improve understanding of error handling for required options

* Refactor argument parsing to use spread syntax for improved readability in multiple test files

* Refactor command argument parsing and update project configurations

- Updated tasks.json to include a new build task for the AzureMcp solution.
- Changed target framework from net9.0 to net10.0 in Directory.Build.props.
- Removed unnecessary package reference to System.Linq.AsyncEnumerable in Microsoft.Mcp.Core.csproj.
- Modified test files to directly parse command arguments instead of using ArgSplitter, simplifying the argument handling.
- Deleted unused ArgBuilder and ArgSplitter utility classes from test utilities.
- Ensured all relevant tests were updated to reflect the new argument parsing method.

* Add readonly default

* Refactor option parsing and default value handling for improved clarity and consistency

* Refactor command validation and option handling for improved clarity and consistency; add new BestPracticesOptions class and enhance task definitions in tasks.json

* Update TODO comments to use consistent naming for better traceability

* Refactor tasks.json to remove duplicate build tasks for improved clarity and maintainability; add CommandResultExtensions class for enhanced command option handling

* Refactor option retrieval methods for consistency across command validation; replace TryGetOptionValue with TryGetValue in multiple command files

* Refactor ExecuteAsync test to use array for command line arguments; remove unnecessary command line parsing logic

* Update package versions and refactor command test implementations for consistency

* Fix formatting in test methods by removing trailing spaces in comments

* Refactor tests to remove skipping and address TODOs for improved test coverage

* Refactor command option binding to use GetValueOrDefault for improved null handling

- Updated various command classes to replace GetValue with GetValueOrDefault for better handling of optional parameters.
- Introduced ParseResultExtensions to encapsulate the new GetValueOrDefault method.
- Adjusted error handling in command execution to utilize the new exception handling structure.
- Added NameNormalization helper class for consistent option name normalization.

* Implement CommandValidationException for structured validation handling and refactor error responses in BaseCommand

* Refactor using directives for consistency across command files

* Refactor firewall rule command classes for improved consistency and error handling

* Enhance RetryPolicyOptions to track explicitly provided settings and update BaseAzureService to apply them conditionally

* Refactor RetryPolicyOptions equality tests for improved clarity and add new tests for flag handling

* Refactor ServiceOptionDefinitions to use 'Required' and 'Hidden' properties for clarity and update DefaultValueFactory syntax

* Update CallToolAsync method calls in CommandTests to use the correct command prefix for best practices retrieval

* Update target framework and package versions for consistency and compatibility

* Fix formatting in ContainerCreateCommandTests for improved readability

+++ tools/Azure.Mcp.Tools.Quota/src/Services/Util/Usage/ComputeUsageChecker.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using Azure.Core;
-using Azure.ResourceManager;
 using Azure.ResourceManager.Compute;
 using Azure.ResourceManager.Compute.Models;
 using Microsoft.Extensions.Logging;
@@ -24,7 +23,7 @@ public class ComputeUsageChecker(TokenCredential credential, string subscription
                 result.Add(new UsageInfo(
                     Name: item.Name?.LocalizedValue ?? item.Name?.Value ?? string.Empty,
                     Limit: (int)item.Limit,
-                    Used: (int)item.CurrentValue,
+                    Used: item.CurrentValue,
                     Unit: item.Unit.ToString()
                 ));
             }
